refact: PiPEnabledPlayerViewController 위치 변경
