+++ Media/Infrastructure/VideoPlayer/VideoPlayerService.swift
@@ -306,3 +306,16 @@ extension DefaultVideoPlayerService: VideoPlayerService {
     }
 
 }
+
+final class PiPEnabledPlayerViewController: AVPlayerViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // PiP 기능 활성화 (영상이 작은 창으로 재생되도록 허용)
+        self.allowsPictureInPicturePlayback = true
+        // 사용자가 별도로 PiP 버튼을 누르지 않아도
+        // 앱이 백그라운드로 전환될 때 자동으로 PiP 모드를 시작할 수 있도록 허용
+        self.canStartPictureInPictureAutomaticallyFromInline = true
+    }
+}
